+++ CHANGELOG.md
@@ -7,10 +7,11 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.0.32
 
-2023 Sep 28 [in progress]
+2023 Oct 4 [in progress]
 
 * Added exception catch to GL string get.
 * Added Waveform widget "auto" marker statistics text position as default  #224
+* Improved Waveform widget autoranging  #228
 
 
 ## 1.0.31